ml/requirements.txt
tensorflow-cpu
flask
pandas
numpy
joblib
flask_cors
sklearn
gunicorn
waitress